requirements.txt
fastapi
uvicorn
pydantic
requests 
openai-whisper
pytube 

python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv

PyJWT